Move domain event logic to deprecated DomainEventTestUtils
Move domain event logic to deprecated DomainEventTestUtils

#3425
