+++ eventsourcing/src/test/java/org/axonframework/eventsourcing/eventstore/jdbc/JdbcEventStorageEngineTest.java
@@ -53,7 +53,7 @@ import java.util.stream.LongStream;
 import java.util.stream.Stream;
 
 import static java.util.stream.Collectors.toList;
-import static org.axonframework.eventhandling.EventTestUtils.*;
+import static org.axonframework.eventhandling.DomainEventTestUtils.*;
 import static org.junit.jupiter.api.Assertions.*;
 import static org.mockito.ArgumentMatchers.*;
 import static org.mockito.Mockito.*;
