+++ integrationtests/src/test/java/org/axonframework/integrationtests/eventsourcing/eventstore/benchmark/AbstractEventStoreBenchmark.java
@@ -49,7 +49,7 @@ import java.util.concurrent.TimeUnit;
 import java.util.stream.Collectors;
 import java.util.stream.IntStream;
 
-import static org.axonframework.eventhandling.EventTestUtils.createDomainEvent;
+import static org.axonframework.eventhandling.DomainEventTestUtils.createDomainEvent;
 
 /**
  * @author Rene de Waele
