+++ messaging/src/test/java/org/axonframework/eventhandling/ConcludesBatchParameterResolverFactoryTest.java
@@ -26,8 +26,8 @@ import org.junit.jupiter.api.*;
 import java.lang.reflect.Method;
 import java.util.List;
 
+import static org.axonframework.eventhandling.DomainEventTestUtils.createDomainEvents;
 import static org.axonframework.eventhandling.EventTestUtils.asEventMessage;
-import static org.axonframework.eventhandling.EventTestUtils.createDomainEvents;
 import static org.axonframework.messaging.unitofwork.StubProcessingContext.forMessage;
 import static org.junit.jupiter.api.Assertions.*;
