+++ messaging/src/test/java/org/axonframework/eventhandling/DirectEventProcessingStrategyTest.java
@@ -21,7 +21,7 @@ import org.junit.jupiter.api.*;
 import java.util.List;
 import java.util.function.Consumer;
 
-import static org.axonframework.eventhandling.EventTestUtils.createDomainEvents;
+import static org.axonframework.eventhandling.DomainEventTestUtils.createDomainEvents;
 import static org.mockito.Mockito.*;
 
 /**
