+++ messaging/src/test/java/org/axonframework/eventhandling/EventTestUtils.java
@@ -96,73 +96,4 @@ public abstract class EventTestUtils {
                 GenericEventMessage.clock.instant()
         );
     }
-
-    // Deciding not to document the methods below until we actually need a non-DomainEventMessage version.
-
-    public static List<DomainEventMessage<?>> createDomainEvents(int numberOfEvents) {
-        return createDomainEvents(() -> AGGREGATE, numberOfEvents);
-    }
-
-    public static List<DomainEventMessage<?>> createDomainEvents(Supplier<String> aggregateId, int numberOfEvents) {
-        return IntStream.range(0, numberOfEvents)
-                        .mapToObj(sequenceNumber -> createDomainEvent(AGGREGATE_TYPE,
-                                                                      IdentifierFactory.getInstance()
-                                                                                       .generateIdentifier(),
-                                                                      aggregateId.get(),
-                                                                      sequenceNumber,
-                                                                      PAYLOAD + sequenceNumber,
-                                                                      METADATA))
-                        .collect(Collectors.toList());
-    }
-
-    public static DomainEventMessage<String> createDomainEvent() {
-        return createDomainEvent(0);
-    }
-
-    public static DomainEventMessage<String> createDomainEvent(long sequenceNumber) {
-        return createDomainEvent(AGGREGATE, sequenceNumber);
-    }
-
-    public static DomainEventMessage<String> createDomainEvent(long sequenceNumber, Instant timestamp) {
-        return new GenericDomainEventMessage<>(
-                AGGREGATE_TYPE, AGGREGATE, sequenceNumber,
-                IdentifierFactory.getInstance().generateIdentifier(), TYPE,
-                PAYLOAD, METADATA, timestamp
-        );
-    }
-
-    public static DomainEventMessage<String> createDomainEvent(String aggregateId, long sequenceNumber) {
-        return createDomainEvent(aggregateId, sequenceNumber, PAYLOAD);
-    }
-
-    public static DomainEventMessage<String> createDomainEvent(String aggregateId, long sequenceNumber,
-                                                               String payload) {
-        return createDomainEvent(AGGREGATE_TYPE,
-                                 IdentifierFactory.getInstance().generateIdentifier(),
-                                 aggregateId,
-                                 sequenceNumber,
-                                 payload,
-                                 METADATA);
-    }
-
-    public static DomainEventMessage<String> createDomainEvent(String eventId, String aggregateId,
-                                                               long sequenceNumber) {
-        return createDomainEvent(AGGREGATE_TYPE, eventId, aggregateId, sequenceNumber, PAYLOAD, METADATA);
-    }
-
-    public static DomainEventMessage<String> createDomainEvent(String type,
-                                                               String eventId,
-                                                               String aggregateId,
-                                                               long sequenceNumber,
-                                                               String payload,
-                                                               MetaData metaData) {
-        return new GenericDomainEventMessage<>(type,
-                                               aggregateId,
-                                               sequenceNumber,
-                                               eventId,
-                                               TYPE,
-                                               payload,
-                                               metaData,
-                                               GenericDomainEventMessage.clock.instant());
-    }
 }
